moved jsonschema diff to pytest-simcore
